NLP/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
groq>=0.4.0
python-dotenv>=1.0.0
